Renamed, removed, and moved request / review files
Removed excess folders, renamed review-request-handlers to comply with REST naming scheme.

Modified .gitignore, adding .vscode.
