+++ server/routes/review.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const reviewModel = require("../review_request_handlers/review/review");
+const reviewModel = require("../review-request-handlers/review");
 
 // import model User for searching
 const User = require("../models/user");
@@ -19,25 +19,20 @@ router.post("/createReview", async (req, res) => {
       messageText: req.body.messageText,
     };
 
-    await reviewModel.createReview(user, data, () => {
-      res.status(201).json({ message: "Complete" });
-    });
+    await reviewModel.createReview(user, data, () => {});
   } catch {
-    res.status(500).json({ message: "Internal Server Error" });
+    console.log("There was an error in creating review");
   }
 });
 
 // gets all relevant reviews
 router.post("/getReviews", async (req, res) => {
   try {
-    const email = req.body.email;
-    const user = await User.findOne({ email: email });
+    // Need to implement user auth middleware here and search off the returned id
 
     const reviews = await Review.find({ userId: user._id });
-
-    res.status(201).json({ result: reviews });
   } catch {
-    res.status(500).json({ message: "Internal Server Error" });
+    console.log("There was an error getting reviews");
   }
 });
